### 2022-12-29 #5037 @editorialbot
[PRE REVIEW]: Atramhasis: An online SKOS editor
**Submitting author:** @koenedaele (<a href="http://orcid.org/0000-0002-8153-2978">Koen Van Daele</a>)
**Repository:** https://github.com/OnroerendErfgoed/atramhasis
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.2.0
