### 2022-12-29 #5037 @danielskatz
👋 @koenedaele - I'm the track editor for this submission to JOSS. Given the generality of JOSS, it would be useful for you to explain/define SKOS and RDF in the paper. I would also suggest a change in the title from "webbased" (which I don't think is English) to "web-based". A couple of other minor comments are to change "Github" to "GitHub" in the acknowledgements, and for the last references, to remove the "doi:" part of the DOI.
