### 2022-12-29 #5037 @danielskatz
Please feel free to make changes to your .bib and/or .md files, then use the commands `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
