+++ paper/paper.bib
@@ -21,7 +21,7 @@
 	volume = {37--38},
 	pages = {184--206},
 	issn = {1570-8268},
-	doi = {doi:10.1016/j.websem.2016.03.003},
+	doi = {10.1016/j.websem.2016.03.003},
 	url = {http://linkeddatafragments.org/publications/jws2016.pdf},
 }
 
