### 2023-01-02 #5037 @koenedaele
Thank you @danielskatz . I've fixed the minor issues, I'll have a look at explaining SKOS and RDF.

For reviewers from my own network of contacts I would suggest Pieter Colpaert (pietercolpaert), Miel Vander Sande (mielvds) or Brecht Van de Vyvere (brechtvdv) although I haven't contacted any of them directly about this.

From the list of reviewers you've provided I would suggest akritiko or gaurav.
