### 2023-01-02 #5037 @pietercolpaert
> wave @pietercolpaert - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I believe I have a conflict of interest. My research team is often being funded by Flemish government agencies and I often use Koen Van Daele’s work on Atramhasis as a good practice. While I would enjoy reviewing the paper, I think the paper could benefit from a more impartial reviewer.
