### 2023-01-02 #5037 @pietercolpaert
Yes, I read it and saw a conflict in the fact that my team is also being funded by Flemish governmental agencies. As potential alternatives, I’d suggest:
 - Seth van Hooland
 - Giorgia Lodi
 - Oscar Corcho
 - Maria Poveda
 - Pascal Hitzler
 - David Chaves
 - Anastasia Dimou
 - Sven Lieber
