### 2023-01-02 02b4606
Update paper with more information on SKOS and RDF.
