+++ paper/paper.bib
@@ -35,6 +35,17 @@
   	year = 2008
 }
 
+@misc{rdf:2014,
+	abstract = {The Resource Description Framework (RDF) is a language for representing information about resources in the World Wide Web. This Primer is designed to provide the reader with the basic knowledge required to effectively use RDF. It introduces the basic concepts of RDF and describes its XML syntax. It describes how to define RDF vocabularies using the RDF Vocabulary Description Language, and gives an overview of some deployed RDF applications. It also describes the content and purpose of other RDF specification documents.},
+  	editor = {Schreiber, Guus and Raimond, Yves},
+  	month = {June},
+  	publisher = {World Wide Web Consortium},
+  	series = {W3C Recommendation},
+  	title = {{RDF 1.1 Primer}},
+	url = {https://www.w3.org/TR/rdf11-primer/}
+  	year = 2014
+}
+
 @misc{aat,
 	author={{Getty Vocabulary Program}},
 	title={Art and Archtitecture Thesaurus},
