+++ paper/paper.bib
@@ -42,7 +42,7 @@
   	publisher = {World Wide Web Consortium},
   	series = {W3C Recommendation},
   	title = {{RDF 1.1 Primer}},
-	url = {https://www.w3.org/TR/rdf11-primer/}
+	url = {https://www.w3.org/TR/rdf11-primer/},
   	year = 2014
 }
 
