+++ paper/paper.bib
@@ -1,4 +1,3 @@
-
 @article{Mortier:2017,
 	title = {Calling it what it is. {Thesauri} in the {Flanders} {Heritage} {Agency}: {History}, {Importance}, {Use} and {Technological} {Advances}},
 	volume = {IV-2/W2},
@@ -23,6 +22,16 @@
 	issn = {1570-8268},
 	doi = {10.1016/j.websem.2016.03.003},
 	url = {http://linkeddatafragments.org/publications/jws2016.pdf},
+}
+
+@book{Harpring:2010,
+  	abstract = {An online publication that defines the characteristics, scope, and uses of controlled vocabularies for art and cultural materials, and explains how vocabularies should be integrated in cataloging systems and utilized for indexing and retrieval.},
+  	author = {Harpring, Patricia},
+  	isbn = {978-1-60606-026-1},
+  	publisher = {Getty Research Institute},
+  	title = {Introduction to Controlled Vocabularies: Terminology for Art, Architecture, and Other Cultural Works},
+  	url = {https://www.getty.edu/research/publications/electronic_publications/intro_controlled_vocab/},
+  	year = 2010
 }
 
 @misc{skos:2008,
@@ -64,4 +73,3 @@
   doi          = {10.5281/zenodo.6984378},
   url          = {https://doi.org/10.5281/zenodo.6984378}
 }
-
