### 2023-01-02 #5037 @koenedaele
@danielskatz I've added some more information on SKOS and RDF and an extra reference to the RDF primer and a publication with more information on the use of controlled vocabularies for cultural heritage. Given the focus of JOSS on software and the limited number of words allowed I've tried to keep it brief. Do let me know if you want me to expand on this section.
