### 2023-01-02 f197ea9
Upgrade skosprovider_rdf. (#742)
* Upgrade skosprovider_rdf.

Issue #741

* Fix action

Co-authored-by: Koen Van Daele <koen_van_daele@telenet.be>
