### 2023-01-02 #5037 @koenedaele
@danielskatz I've had a look at @pietercolpaert 's list. Some do not seem to have a GH account or have no knowledge of Python as far as I can see.
I think Sven Lieber (SvenLieber) might be a good candidate since he knows Linked Data, Cultural Heritage and has some Python repos. David Chaves (I think this is dachafra) seems to have knowledge of Linked Data and some Python experience.
