### 2023-01-04 8424cc7
Switch development to 1.3.0
