+++ docs/source/conf.py
@@ -49,16 +49,16 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'Atramhasis'
-copyright = '2014-2021, Flanders Heritage Agency'
+copyright = '2014-2023, Flanders Heritage Agency'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = '1.1'
+version = '1.3'
 # The full version, including alpha/beta/rc tags.
-release = '1.1.0'
+release = '1.3.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
