+++ CITATION.cff
@@ -81,5 +81,5 @@ keywords:
   - vocabulary
   - python
 license: GPL-3.0
-version: 1.2.0
-date-released: '2022-10-19'
+version: 1.3.0
+date-released: '2023-01-??'
