### 2023-01-04 38ddc86
Limit requirements for a functioning atramhasis.
Requirements will be upgraded in a future major release.
These limitations require no changes to the current code.

Issue #749
