### 2023-01-04 abdfed6
Limit requirements for a functioning atramhasis. (#750)
Requirements will be upgraded in a future major release.
These limitations require no changes to the current code.

Issue #749
