+++ CITATION.cff
@@ -82,4 +82,4 @@ keywords:
   - python
 license: GPL-3.0
 version: 1.3.0
-date-released: '2023-01-??'
+date-released: '2023-01-04'
