### 2023-01-05 9db9b0e
Remove universal flag for wheel creation.
Only python 3 supported.

Issue #752
