+++ setup.cfg
@@ -1,6 +1,3 @@
-[bdist_wheel]
-universal=1
-
 [compile_catalog]
 directory = atramhasis/locale
 domain = atramhasis
