### 2023-01-05 #5037 @SvenLieber
@danielskatz yes I would be available to review this submission for JOSS in the following 1-2 weeks.
