### 2023-01-05 #5040 @editorialbot
[REVIEW]: Atramhasis: A webbased SKOS editor
**Submitting author:** @koenedaele (<a href="http://orcid.org/0000-0002-8153-2978">Koen Van Daele</a>)
**Repository:** https://github.com/OnroerendErfgoed/atramhasis
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.3.2
