+++ fixtures/materiaal_data.py
@@ -1,981 +1,1388 @@
 MATERIAALDATA = [
-    {   'broader': [],
-        'id': 1,
-        'labels': [   {   'label': 'aardewerk',
-                           'language': 'nl-BE',
-                           'type': 'prefLabel'},
-                       {   'label': 'ceramiek',
-                           'language': 'nl-BE',
-                           'type': 'altLabel'},
-                       {   'label': 'keramiek',
-                           'language': 'nl-BE',
-                           'type': 'altLabel'}],
-        'matches': {   'broad': [],
-                        'close': [],
-                        'exact': [],
-                        'narrow': [],
-                        'related': []},
-        'member_of': [],
-        'narrower': [2, 5, 6],
-        'notes': [   {   'language': 'nl-BE',
-                          'markup': None,
-                          'note': 'Aardewerk of keramiek is gebakken klei. Aardewerk wordt gebruikt voor het vervaardigen van vaatwerk, bouwmaterialen, beeldjes, enz\u2026',
-                          'type': 'scopeNote'}],
-        'related': [],
-        'sources': [   {   'citation': 'Nederlandse Art &amp; Architecture Thesaurus, <a href="http://browser.aat-ned.nl/">http://browser.aat-ned.nl/</a>',
-                            'markup': 'HTML'}],
-        'subordinate_arrays': [],
-        'type': 'concept',
-        'uri': 'https://id.erfgoed.net/thesauri/materialen/1'},
-    {   'broader': [8],
-        'id': 48,
-        'labels': [   {   'label': 'aluminium',
-                           'language': 'nl-BE',
-                           'type': 'prefLabel'}],
-        'matches': {   'broad': [],
-                        'close': [],
-                        'exact': [],
-                        'narrow': [],
-                        'related': []},
-        'member_of': [],
-        'narrower': [],
-        'notes': [   {   'language': 'nl-BE',
-                          'markup': None,
-                          'note': 'Aluminium is een metaal uit de derde groep van het periodiek systeem (chemisch element Al). Het is dof zilverachtig van kleur en zeer licht.',
-                          'type': 'scopeNote'}],
-        'related': [],
-        'sources': [   {   'citation': 'Van Dale groot woordenboek van de Nederlandse taal. 14de editie, Utrecht/Antwerpen, 2005',
-                            'markup': None}],
-        'subordinate_arrays': [],
-        'type': 'concept',
-        'uri': 'https://id.erfgoed.net/thesauri/materialen/48'},
-    {   'broader': [32],
-        'id': 33,
-        'labels': [   {   'label': 'amber',
-                           'language': 'nl-BE',
-                           'type': 'prefLabel'},
-                       {   'label': 'barnsteen',
-                           'language': 'nl-BE',
-                           'type': 'altLabel'}],
-        'matches': {   'broad': [],
-                        'close': [],
-                        'exact': [],
-                        'narrow': [],
-                        'related': []},
-        'member_of': [],
-        'narrower': [],
-        'notes': [   {   'language': 'nl-BE',
-                          'markup': None,
-                          'note': 'Amber is een fossiele hars afkomstig uit naaldbomen en wordt gebruikt voor het maken van kralen, hangers, knopen en dergelijke meer.',
-                          'type': 'scopeNote'}],
-        'related': [],
-        'sources': [{   'citation': 'VIOE', 'markup': None}],
-        'subordinate_arrays': [],
-        'type': 'concept',
-        'uri': 'https://id.erfgoed.net/thesauri/materialen/33'},
-    {   'broader': [],
-        'id': 38,
-        'labels': [   {   'label': 'beton',
-                           'language': 'nl-BE',
[…]

[…]
