### 2023-01-17 #5040 @gaurav
Hi everybody,

Thank you so much for an opportunity to review this manuscript! I got stuck pretty early on in the review process because the [Atramhasis repository](https://github.com/OnroerendErfgoed/atramhasis/) doesn't have installation instructions. I tried following the installation instructions at [your readthedocs.io page](https://atramhasis.readthedocs.io/en/latest/development.html#general-installation) page, but running `initialize_atramhasis_db development.ini` resulted in the following error:

```
raise VersionConflict(dist, req).with_context(dependent_req)
pkg_resources.ContextualVersionConflict: (MarkupSafe 2.0.1 (/Users/gaurav/Development/open-source/atramhasis/venv/lib/python3.9/site-packages), Requirement.parse('MarkupSafe>=2.1.1'), {'werkzeug'})
```

Could you please copy the basic instructions for starting Atramhasis into the repository's README file? I'll keep trying to get Atramhasis to run, but if you have any insights into why running that command resulted in a version conflict, please let me know!

Here are my comments on the GitHub repository so far:
- [ ] JOSS asks: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". While you clearly meet the first requirement in clearly described in CONTRIBUTING.md, I think a few sentences encouraging users to file issues or seek support at https://github.com/OnroerendErfgoed/atramhasis/issues would be nice.

Here are my comments on the manuscript so far:
- [ ] JOSS wanted me to check whether "the full list of paper authors seem appropriate and complete". I noticed in your repository's [contributors list](https://github.com/OnroerendErfgoed/atramhasis/graphs/contributors) that Maarten Taeymans contributed significantly to this software but isn't credited as a co-author or in the acknowledgements. I will accept your claim if you say that they didn't contribute significantly to the software as it currently exists, but I just wanted to note this in case it was an oversight on your part.
- Summary
  - [ ] Unless there's a specific reason for citing [the 2008-08-29 version of SKOS](https://www.w3.org/TR/2008/WD-skos-reference-20080829/), you might as well cite the current version of SKOS, which is [2009-08-18 version of SKOS](https://www.w3.org/TR/2009/REC-skos-reference-20090818/).
- Statement of need
  - [ ] Your first paragraph does a good job of explaining why you built Atramhasis. However, I think a sentence or two on how it compares to other similar tools (such as [Skosmos](https://skosmos.org/)) and what distinguishes it from them would be useful here.
  - [ ] It might be useful to make explicit who your target audience is: do you have a sense of the size of the thesaurus or the number of users that Atramhasis can support? Can a single Atramhasis instance support multiple thesauri? Do you anticipate thesaurus creators running Atramhasis locally on their computer when they want to edit a thesaurus, or do you anticipate an organization setting up an instance that multiple users can use?
  - [ ] "Concept uris" -- I've generally seen "uris" capitalized as "URIs", and I think that makes sense here.
  - [ ] It's "JSON-LD", not "JSON/LD". I think a citation to the W3 spec at https://www.w3.org/TR/json-ld11/ would be useful too.
  - [ ] I understand that you don't have enough space here to describe the Linked Data Fragments (LDF) server in any detail, but I would include a link to your [documentation on this functionality]
(https://atramhasis.readthedocs.io/en/latest/development.html#running-a-linked-data-fragments-server) in your manuscript. If there is an LDF endpoint at https://thesaurus.onroerenderfgoed.be/, I think that would be useful to link to as well!
- Minor changes, not required but suggested
  - I notice that you reference figures by saying e.g. "edit data in a normal web admin interface \autoref{fig:editingairfields}.". This appears as "... interface Figure 2." in the generated PDF. I would recommend either saying "in a normal web interface, as seen in \autoref{fig:editingairfields}." or "in a normal web interface (\autoref{fig:editingairfields})." to make it a bit more readable.
