### 2023-01-17 #5040 @koenedaele
Thank you for agreeing to review our software! For installation, I would recommend creating a demo environment via https://atramhasis.readthedocs.io/en/latest/demo.html. This should work with python 3.8, 3.9 and 3.10.

We'll check into that version conflict. Which Python version are you running?
