### 2023-01-17 OnroerendErfgoed/atramhasis#755 @koenedaele
Can't install development version
In openjournals/joss-reviews#5040 @gaurav has noticed that:

I got stuck pretty early on in the review process because the [Atramhasis repository](https://github.com/OnroerendErfgoed/atramhasis/) doesn't have installation instructions. I tried following the installation instructions at [your readthedocs.io page](https://atramhasis.readthedocs.io/en/latest/development.html#general-installation) page, but running initialize_atramhasis_db development.ini resulted in the following error:

`
raise VersionConflict(dist, req).with_context(dependent_req)
pkg_resources.ContextualVersionConflict: (MarkupSafe 2.0.1 (/Users/gaurav/Development/open-source/atramhasis/venv/lib/python3.9/site-packages), Requirement.parse('MarkupSafe>=2.1.1'), {'werkzeug'})
`

Normally things should work for Python 3.9.
