+++ paper/paper.bib
@@ -40,7 +40,7 @@
   	howpublished = {World Wide Web Consortium, Working Draft WD-skos-reference-20080829},
   	month = {August},
   	title = {SKOS Simple Knowledge Organization System Reference},
-  	url = {http://www.w3.org/TR/2008/WD-skos-reference-20080829},
+	url = {https://www.w3.org/TR/2009/REC-skos-reference-20090818/},
   	year = 2008
 }
 
