### 2023-01-17 7981984
Write uris as URIs
