### 2023-01-17 OnroerendErfgoed/atramhasis#755 @Wim-De-Clercq
That's https://github.com/OnroerendErfgoed/atramhasis/issues/747
Because we don't support the latest jinja2, we also do not support the latest markupsafe.
