+++ paper/paper.bib
@@ -55,6 +55,17 @@
   	year = 2014
 }
 
+@misc{jsonld:2020
+	abstract = {JSON is a useful data serialization and messaging format. This specification defines JSON-LD 1.1, a JSON-based format to serialize Linked Data. The syntax is designed to easily integrate into deployed systems that already use JSON, and provides a smooth upgrade path from JSON to JSON-LD. It is primarily intended to be a way to use Linked Data in Web-based programming environments, to build interoperable Web services, and to store Linked Data in JSON-based storage engines.},
+	editor = {Kellogg, Greg and Champin, Pierre-Antoine and Longley, Dave},
+	url = {https://www.w3.org/TR/2020/REC-json-ld11-20200716/},
+	title = {{JSON-LD 1.1}}
+	year = 2020,
+	month = {July},
+  	series = {W3C Recommendation},
+  	publisher = {World Wide Web Consortium},
+}
+
 @misc{aat,
 	author={{Getty Vocabulary Program}},
 	title={Art and Archtitecture Thesaurus},
