### 2023-01-17 c06fd6e
Better reference to figure 2
