### 2023-01-17 #5040 @koenedaele
Update paper with correct citation of SKOS, URIs, correct spelling of JSON-LD and reference to JSON-LD 1.1 and better reference to figure 2.
