### 2023-01-17 OnroerendErfgoed/atramhasis#756 @koenedaele
Update CONTRIBUTING.md
From openjournals/joss-reviews#5040 by @gaurav:

`
JOSS asks: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". While you clearly meet the first requirement in clearly described in CONTRIBUTING.md, I think a few sentences encouraging users to file issues or seek support at https://github.com/OnroerendErfgoed/atramhasis/issues would be nice.
`
