### 2023-01-17 #5040 @koenedaele
@gaurav I've expanded the list of authors of the paper with the people who were involved in development in a significant way, although some of them have left our organisation and were not involved in this paper.
