### 2023-01-18 OnroerendErfgoed/atramhasis#759 @SvenLieber
How to run the cookiecutter demo with fixed dependencies?
Hi,

as part of reviewing this software for the Journal of Open Source Software (https://github.com/openjournals/joss-reviews/issues/5040) I tried to run a demo of the tool, but got some errors which I think might be related to https://github.com/OnroerendErfgoed/atramhasis/issues/749

My question is what the recommended procedure is to get the demo running with fixed dependecies.
The only dependencies I see are

`requirements-dev.txt`
```
# Runtime requirements
--requirement requirements.txt

pyramid-debugtoolbar
```

`requirements.txt`
```
atramhasis
```

### Description
I expected a running demo server, but instead got errors during the installation within the created python virtual environment about dependencies and eventually an `openapi spec validator`-related error when trying to run the server or create an RDF dump. See below for a detailed description.

### Steps to reproduce
I followed the latest documentation on running a demo site with Cookiecutter.

#### Step 1: Install cookiecutter
```
$ mkvirtualenv atramhasis_demo
$ pip install -U cookiecutter
```
First I got an error because `mkvirtualenv` was not installed. Installing it properly turned out to be tricky, so instead I did the following:

```
$ python3 -m venv py-atramhasis-env
$ . py-atramhasis-env/bin/activate
$ pip install -U cookiecutter
```

#### Step 2: generate the demo site with cookiecutter worked without problems

```bash
$ cookiecutter gh:OnroerendErfgoed/atramhasis_demo_cookiecutter
```

#### Step 3: install dependencies

```bash
$ cd atramhasis_demo
$ pip install -r requirements-dev.txt
$ pip install -e .
```

This results in the following errors

```bash
ERROR: openapi-core 0.16.4 has requirement openapi-schema-validator<0.4.0,>=0.3.0, but you'll have openapi-schema-validator 0.2.3 which is incompatible.
ERROR: openapi-core 0.16.4 has requirement openapi-spec-validator<0.6.0,>=0.5.0, but you'll have openapi-spec-validator 0.4.0 which is incompatible.
```

#### Step 4: Setup the database worked without problems

```bash
$ alembic upgrade head
# fill the database with data
$ initialize_atramhasis_db development.ini
```

#### Step 5: Optionally create RDF dump

```bash
$ dump_rdf development.ini
```

This results in the following error:

```bash
from openapi_spec_validator.validation import openapi_spec_validator_proxy
ModuleNotFoundError: No module named 'openapi_spec_validator.validation'
```

#### Step 6: Install front-end dependencies worked without problems

```bash
$ cd atramhasis_demo/static
$ npm install
```

#### Step 7: Start the server

```bash
$ cd ../..
# start server
$ pserve development.ini
```
results in the following error

```bash
from openapi_spec_validator.validation import openapi_spec_validator_proxy
ModuleNotFoundError: No module named 'openapi_spec_validator.validation'
```

### Environment

I work in an Ubuntu 20.04 virtual machine with `Python 3.8.10` and `pip 20.0.2` (in the virtual environment successfully created in the first step).
