### 2023-01-18 8f76176
Update contributing.md file. Refs #756 (#757)
* Update contributing
* Update changelog
