+++ CHANGES.rst
@@ -1,3 +1,8 @@
+1.3.1 (13-03-2023)
+------------------
+
+- Update CONTRIBUTING.md file (#756)
+
 1.3.0 (04-01-2023)
 -------------------
 
