### 2023-01-18 3861aab
Update contributing.md file. Refs #756 (#757)
* Update contributing
* Update changelog
