### 2023-01-18 OnroerendErfgoed/atramhasis#759 @koenedaele
Looks like an issue with Python 3.8. I just ran your steps with Python 3.10 on Ubuntu 22.04 and that works, so I assume it's the Python 3.8 that makes the difference. The openapi libraries have been a hassle. We'll have a look at how we can fix this quickly.

As to the virtualenv, that's the tool I like to use, but there are indeed several different ways of installing a venv. I'll clarify that in the docs.
