### 2023-01-18 OnroerendErfgoed/atramhasis#759 @koenedaele
I'm not sure where openapi-core 0.16.4 is coming from. Normally it should be 0.13.8 which is required by pyramid_openapi3 0.14.3.

When executing these commands you normally get output on what's being installed. Could you post that?

```
$ cd atramhasis_demo
$ pip install -r requirements-dev.txt
$ pip install -e .
```

Or in you current virtualenvironment do a `pip list`
