+++ paper/paper.bib
@@ -34,6 +34,15 @@
   	year = 2010
 }
 
+@inproceedings{Taelman:2018,
+  author    = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
+  title     = {Comunica: a Modular SPARQL Query Engine for the Web},
+  booktitle = {Proceedings of the 17th International Semantic Web Conference},
+  year      = {2018},
+  month     = oct,
+  url       = {https://comunica.github.io/Article-ISWC2018-Resource/}
+}
+
 @misc{skos:2008,
   	abstract = {This document defines the Simple Knowledge Organization System (SKOS), a common data model for sharing and linking knowledge organization systems via the Web. Many knowledge organization systems, such as thesauri, taxonomies, classification schemes and subject heading systems, share a similar structure, and are used in similar applications. SKOS captures much of this similarity and makes it explicit, to enable data and technology sharing across diverse applications. The SKOS data model provides a standard, low-cost migration path for porting existing knowledge organization systems to the Semantic Web. SKOS also provides a light weight, intuitive language for developing and sharing new knowledge organization systems. It may be used on its own, or in combination with formal knowledge representation languages such as the Web Ontology language (OWL). This document is the normative specification of the Simple Knowledge Organization System. It is intended for readers who are involved in the design and implementation of information systems, and who already have a good understanding of Semantic Web technology, especially RDF and OWL.},
   	author = {Miles, Alistair and Bechhofer, Sean},
