### 2023-01-19 OnroerendErfgoed/atramhasis#759 @SvenLieber
Thanks, after installing Python 3.10 and using a virtual environment created from it, **running the demo worked!**

For completeness, here is the output of `pip list` from the Python 3.8 environment:

```bash
Package                   Version   Location
------------------------- --------- ----------------------------------------------------
alembic                   1.9.2
arrow                     1.2.3
atramhasis                1.3.0
atramhasis-demo           0.0       /home/slieber/repos/skos-editor-test/atramhasis_demo
attrs                     22.2.0
Babel                     2.11.0
backports.cached-property 1.0.2
binaryornot               0.4.4
bleach                    5.0.1
CacheControl              0.12.11
cachetools                5.2.1
certifi                   2022.12.7
chardet                   5.1.0
charset-normalizer        3.0.1
click                     8.1.3
colander                  2.0
cookiecutter              2.1.1
decorator                 5.1.1
dogpile.cache             1.1.8
frozendict                2.3.4
greenlet                  2.0.1
html5lib                  1.1
hupper                    1.11
idna                      3.4
importlib-metadata        6.0.0
importlib-resources       5.10.2
iso8601                   1.1.0
isodate                   0.6.1
Jinja2                    2.11.3
jinja2-time               0.2.0
jsonschema                4.17.3
jsonschema-spec           0.1.2
language-tags             1.2.0
lxml                      4.9.2
Mako                      1.2.4
MarkupSafe                2.0.1
more-itertools            9.0.0
msgpack                   1.0.4
openapi-core              0.16.4
openapi-schema-validator  0.2.3
openapi-spec-validator    0.4.0
parse                     1.19.0
PasteDeploy               3.0.1
pathable                  0.4.3
pbr                       5.11.1
pip                       20.0.2
pkg-resources             0.0.0
pkgutil-resolve-name      1.3.10
plaster                   1.1.2
plaster-pastedeploy       1.0.1
Pygments                  2.14.0
PyLD                      2.0.3
pyparsing                 3.0.9
pyramid                   2.0
pyramid-debugtoolbar      4.10
pyramid-jinja2            2.10
pyramid-mako              1.1.0
pyramid-openapi3          0.15.0
pyramid-rewrite           0.2
pyramid-skosprovider      1.1.0
pyramid-tm                2.5
pyrsistent                0.19.3
python-dateutil           2.8.2
python-slugify            7.0.0
pytz                      2022.7.1
PyYAML                    6.0
rdflib                    6.2.0
requests                  2.28.2
rfc3987                   1.3.8
setuptools                44.0.0
six                       1.16.0
skosprovider              1.2.0
skosprovider-getty        1.1.0
skosprovider-rdf          1.3.0
skosprovider-sqlalchemy   1.0.0
SQLAlchemy                1.4.46
stevedore                 4.1.1
text-unidecode            1.3
transaction               3.0.1
translationstring         1.4
typing-extensions         4.4.0
urllib3                   1.26.14
venusian                  3.0.0
waitress                  2.1.2
webencodings              0.5.1
WebOb                     1.8.7
Werkzeug                  2.2.2
zipp                      3.11.0
zope.deprecation          4.4.0
zope.interface            5.5.2
zope.sqlalchemy           1.6
```
