### 2023-01-19 OnroerendErfgoed/atramhasis#760 @SvenLieber
Unclear distinction in the demo between the languages "Flemish" and "Dutch, Flemish, Belgium"
A tiny detail I noticed while testing the demo for https://github.com/openjournals/joss-reviews/issues/5040

### Description

In the following dialog a user might get confused when being offered Flemish twice.

![image](https://user-images.githubusercontent.com/3501171/213426186-7a2bac8d-13c1-4363-80b0-303c0aba7e68.png)

After saving in the UI (and also in a SKOS RDF export), one of the two different Flemish labels has the language code `vls` which is used for West Flemish (according to [Wikipedia](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes)), but this was not clear from the UI in the dialog.

However, this is probably negligible as displayed languages can be changed from the UI (in which also new languages can be added) and this is just a demo environment with some prefilled languages.

### Steps to reproduce

1. Running the demo using cookiecutter
2. Opening the admin interface
3. Editing an existing concept
