### 2023-01-19 OnroerendErfgoed/atramhasis#759 @Wim-De-Clercq
The pip and setuptools versions are too old (those are from 2020), and not good enough to deal with clashing dependencies.

The (snipped) dependency tree you end up with, with the old versions is this:
```
atramhasis==1.3.0
  - openapi-spec-validator [required: 0.4.0, installed: 0.4.0]
  - pyramid-openapi3 [required: Any, installed: 0.15.0]
    - openapi-core [required: >=0.16.1, installed: 0.16.5]
      - openapi-spec-validator [required: >=0.5.0,<0.6.0, installed: 0.4.0]
```
Which is clearly wrong because the `openapi-spec-validator` requirement from `openapi-core` does not match with what was installed.

With a modern `pip` and `setuptools` the dependency resolver is smart enough to realise this and use older versions which do match.
If you install the dependencies after you do a `pip install --upgrade pip setuptools` you should end up with
```
atramhasis==1.3.0
  - openapi-spec-validator [required: 0.4.0, installed: 0.4.0]
  - pyramid-openapi3 [required: Any, installed: 0.14.3]
    - openapi-core [required: >=0.13.4,<0.14, installed: 0.13.8]
      - openapi-spec-validator [required: Any, installed: 0.4.0]
```
which is valid.
