### 2023-01-19 OnroerendErfgoed/atramhasis#760 @koenedaele
Yes, it is confusing. I probably should have picked another example language, but I was born in West-Flanders ;-)

The languages get validated against the IANA language registry and that's also confusing:
```
[
  {
                "Type": "language",
                "Subtag": "vls",
                "Description": [
                        "Vlaams"
                ],
                "Added": "2009-07-29"
        },
       {
                "Type": "language",
                "Subtag": "nl",
                "Description": [
                        "Dutch",
                        "Flemish"
                ],
                "Added": "2005-10-16",
                "Suppress-Script": "Latn"
        },
]
```

It's rather weird that vls is Vlaams, but nl is Flemish.

I'm also not sure if IANA vls is actually West-Flemish or what linguists would call Flemish (with sub languages West-Flemish, East-Flemish, French-Flemish and Zeelandic Flemish).

Anyway, it might be a good idea to change the label of vls in the demo to Vlaams, that way it's consistent with the official IANA declaration and it is what you would get if you imported a rdf language string such as "Toekter"@vls. But still confusing. So, maybe it would actually be best to update those dialogs (adding labels and notes) to show not just the Language label but also the code. We normally don't show identifiers but in this case the identifiers are pretty well known strings so it shouldn't be a problem.
