### 2023-01-20 OnroerendErfgoed/atramhasis#763 @koenedaele
Trigger reload of list of concepts on adding a concept
As noted by @SvenLieber in openjournals/joss-reviews#5040:

> After adding a new concept to a concept scheme, I noticed that the list of concepts on the right was not updated. A reload was necessary. This is a minor issue, but I would have expected that some sort of event gets triggered and the list will be updated with the newly added concept. An impatient user might create a concept several times or may contact the support for such kind of things.

@JonasGe or @cedrikv   I know it can be a bit tricky to implement, but I assume we can trigger a reload after the POST to the server is complete since these are very small records that update quickly?
