### 2023-01-23 OnroerendErfgoed/atramhasis#763 @cedrikv
our usual solution is to trigger a refresh after a delay, should be a quick fix
