+++ requirements.txt
@@ -12,7 +12,7 @@ pyramid_skosprovider==1.1.0
 skosprovider_rdf==1.3.0
 skosprovider_getty==1.1.0
 
-language-tags==1.1.0
+language-tags==1.2.0
 
 # database
 sqlalchemy==1.4.28
