+++ development.ini
@@ -72,8 +72,19 @@ twitter.site =
 # '127.0.0.1' and '::1'.
 # debugtoolbar.hosts = 127.0.0.1 ::1
 
-# Enter your Google Analytics Web Property ID
-# ga.tracker_key = UA-XXXX-Y
+# Enter your tracking snippet
+tracking_snippet = <!-- Google Analytics -->
+    <script type="text/javascript">
+    (function(i,s,o,g,r,a,m){i["GoogleAnalyticsObject"]=r;i[r]=i[r]||function(){
+    (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
+    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+    })(window,document,"script","//www.google-analytics.com/analytics.js","ga");
+
+    ga("create", "some_key", "auto");
+    ga("set", "forceSSL", true);
+    ga("send", "pageview");
+    </script>
+    <!-- End Google Analytics -->
 
 layout.focus_conceptschemes =
   HERITAGETYPE
