### 2023-02-01 #5040 @koenedaele
@danielskatz I'm still working on this. I've linked a ticket in the Atramhasis repo that contains the requested changes so I can track them. Some of them were quite simple, some of the requests on more information require more work. I will ask for a new review once I have a version that I'm reasonably happy with. My main worry is that answering all requests for extra information pushes the paper beyond the normal size limits for joss. But I'm first writing those parts and we can see if parts need to be cut again.
