+++ paper/paper.bib
@@ -24,6 +24,20 @@
 	url = {http://linkeddatafragments.org/publications/jws2016.pdf},
 }
 
+@article{VanDaele:2015,
+	title = {Data-driven systems and system-driven data: the story of the {Flanders} {Heritage} {Inventory} (1995-2015)},
+	volume = {II-5/W3},
+	issn = {2194-9042},
+	url = {http://www.isprs-ann-photogramm-remote-sens-spatial-inf-sci.net/II-5-W3/323/2015/isprsannals-II-5-W3-323-2015.pdf},
+	doi = {10.5194/isprsannals-II-5-W3-323-2015},
+	language = {en},
+	journal = {ISPRS Annals of the Photogrammetry, Remote Sensing and Spatial Information Sciences},
+	author = {Van Daele, Koen and Meganck, Leen and Mortier, Sophie},
+	year = {2015},
+	note = {25th International CIPA Symposium 2015, 31 August – 04 September 2015, Taipei, Taiwan},
+	pages = {323--327},
+}
+
 @book{Harpring:2010,
   	abstract = {An online publication that defines the characteristics, scope, and uses of controlled vocabularies for art and cultural materials, and explains how vocabularies should be integrated in cataloging systems and utilized for indexing and retrieval.},
   	author = {Harpring, Patricia},
@@ -35,12 +49,36 @@
 }
 
 @inproceedings{Taelman:2018,
-  author    = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
-  title     = {Comunica: a Modular SPARQL Query Engine for the Web},
-  booktitle = {Proceedings of the 17th International Semantic Web Conference},
-  year      = {2018},
-  month     = oct,
-  url       = {https://comunica.github.io/Article-ISWC2018-Resource/}
+  	author    = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
+  	title     = {Comunica: a Modular SPARQL Query Engine for the Web},
+  	booktitle = {Proceedings of the 17th International Semantic Web Conference},
+  	year      = {2018},
+  	month     = oct,
+  	url       = {https://comunica.github.io/Article-ISWC2018-Resource/}
+}
+
+@techreport{deHaan:2021,
+	doi = {10.55465/rsdd4339},
+	url = {https://doi.org/10.55465/rsdd4339},
+	year = 2021,
+	month = {sep},
+	institution = {agentschap Onroerend Erfgoed},
+	author = {Aukje de Haan and Nele Vanmaele},
+	title = {Inventariseren van landschappelijk erfgoed in Vlaanderen. Historiek, methodologie, doelstellingen en resultaten},
+	series = {Onderzoeksrapporten agentschap Onroerend Erfgoed},
+	number = {195}
+}
+
+@techreport{Hooft:2021,
+	doi = {10.55465/mmym4330},
+	url = {https://doi.org/10.55465/mmym4330},
+	year = 2021,
+	month = {sep},
+	institution = {agentschap Onroerend Erfgoed},
+	author = {Elise Hooft},
+	title = {Inventariseren van bouwkundig erfgoed in Vlaanderen. Historiek, methodologie, doelstellingen en resultaten},
+	series = {Onderzoeksrapporten agentschap Onroerend Erfgoed},
+	number = {196}
 }
 
 @misc{skos:2008,
