### 2023-02-03 1227f37
Further rewrite. WIP.
