+++ paper/paper.md
@@ -65,54 +65,64 @@ to concepts in other vocabularies such as the Art and Architecture Thesaurus (AA
 
 ![Airfields described as a SKOS concept.\label{fig:airfields}](atramhasis_screen_airfields.png)
 
-# Statement of need
+# Functional requirements
 
-Atramhasis was written to be a user-friendly open source SKOS editor. First and foremost,
-we wanted a system that adheres to the SKOS standard, yet is useable for users without
-knowledge of SKOS or RDF. For a typical user, it had to feel as if they were consulting 
-a normal website, as opposed to a RDF vocabulary since the latter can feel rather daunting 
-to non-technical users. This not only applies to users consulting the thesauri, but also 
-to those editing them. Our thesaurus editors are not IT-specialists, but researchers in the
-field of cultural heritage. While a general understanding of thesauri is within their grasp, 
-the technicalities of RDF are not. Thus, editors in Atramhasis do not write RDF statements,
-but edit data in a normal web admin interface, as seen in \autoref{fig:editingairfields}. 
+Atramhasis was written to be a user-friendly open source SKOS editor. First and 
+foremost, we wanted a system that adheres to the SKOS standard, yet is useable 
+for users without knowledge of SKOS or RDF. For a typical user, it had to feel 
+as if they were consulting a normal website, as opposed to a RDF vocabulary 
+since the latter can feel rather daunting to non-technical users. This not only 
+applies to users consulting the thesauri, but also to those editing them. Our 
+thesaurus editors are not IT- or linked data specialists, but domain experts, 
+researchers and other specialists in the field of cultural heritage. While a 
+general understanding of thesauri is within their grasp, the technicalities of 
+RDF are not. Thus, editors in Atramhasis do not write RDF statements, but edit 
+data in a normal web admin interface, as seen in \autoref{fig:editingairfields}. 
 All mapping to RDF and SKOS is done behind the scene, invisible to the editors.
 
 ![Editing the concept of airfields is simple and straightforward.\label{fig:editingairfields}](atramhasis_screen_edit_airfields.png)
 
-The system was conceived as Flanders Heritage's [central platform](https://thesaurus.onroerenderfgoed.be) 
-for publication of internal and regional vocabularies dealing with cultural heritage 
-[@Mortier:2017]. The publication website allows humans to browse, search and 
-consult the vocabularies online in a user-friendly way. Search results can be 
-downloaded in CSV format for further processing. Internal and external systems 
-use the webservices provided by Atramhasis to consult or download vocabularies. 
-Concept URIs are used in indexing data in systems such as the 
+The system was conceived as Flanders Heritage's 
+[central platform](https://thesaurus.onroerenderfgoed.be) for publication of µ
+internal and regional vocabularies dealing with cultural heritage [@Mortier:2017]. 
+The publication website allows humans to browse, search and consult the vocabularies 
+online in a user-friendly way. Search results can be downloaded in CSV format 
+for further processing. Internal and external systems use the webservices provided 
+by Atramhasis to consult or download vocabularies. Concept URIs are used in 
+indexing data in systems such as the 
 [Inventory of Immovable Cultural Heritage](https://inventaris.onroerenderfgoed.be) 
 [@VanDaele:2015; @deHaan:2021; @Hooft:2021] or the 
 [Flanders Heritage Image Database](https://beeldbank.onroerenderfgoed.be).
-This allows users to search those external systems using the provided thesauri (\autoref{fig:searchingairfields}).
-For a typical end-user the thesauri are presented as dropdown lists or specialised 
-widgets that allow navigating the thesaurus from the top concepts along branches to 
-the leafs. For most internal operations with the thesaurus, JSON REST services are used because 
-they are convenient for web developers to design and develop enterprise IT-systems.
[…]
