+++ atramhasis/static/admin/src/app/ui/dialogs/LabelsDialog.js
@@ -39,7 +39,7 @@ define([
         data: this.langList,
         idProperty: 'id',
         labelProperty: 'name'
-      });
+      }, true);
     },
 
     startup: function () {
