### 2023-02-07 938b0cd
#767 op.execute oneline instead of get bind
