+++ atramhasis/alembic/versions/0978347c16f9_rename_flemish_vlaams.py
@@ -16,9 +16,7 @@ revision = '0978347c16f9'
 down_revision = '88214d86a9d1'
 
 def upgrade():
-    connection = op.get_bind()
-    connection.execute("UPDATE language SET name = 'Vlaams' WHERE id = 'vls'")
+    op.execute("UPDATE language SET name = 'Vlaams' WHERE id = 'vls'")
     
 def downgrade():
-    connection = op.get_bind()
-    connection.execute("UPDATE language SET name = 'Flemish' WHERE id = 'vls'")
+    op.execute("UPDATE language SET name = 'Flemish' WHERE id = 'vls'")
