+++ paper/paper.md
@@ -83,7 +83,7 @@ All mapping to RDF and SKOS is done behind the scene, invisible to the editors.
 ![Editing the concept of airfields is simple and straightforward.\label{fig:editingairfields}](atramhasis_screen_edit_airfields.png)
 
 The system was conceived as Flanders Heritage's 
-[central platform](https://thesaurus.onroerenderfgoed.be) for publication of µ
+[central platform](https://thesaurus.onroerenderfgoed.be) for publication of 
 internal and regional vocabularies dealing with cultural heritage [@Mortier:2017]. 
 The publication website allows humans to browse, search and consult the vocabularies 
 online in a user-friendly way. Search results can be downloaded in CSV format 
@@ -98,17 +98,19 @@ For a typical end-user the thesauri are presented as dropdown lists or
 [specialised widgets](https://github.com/OnroerendErfgoed/thesaurus-widget) that 
 allow navigating the thesaurus from the top concepts along branches to the leafs. 
 For most transactions between internal Flanders Heritage systems and the
-the thesaurus system, simple JSON REST services are used. This was deliberately kept 
-close to the implementation standards used in other Flanders Heritage sytems to allow 
-developers used to working in an enterprise IT context to feel comfortable and 
-productive.
+thesaurus system, simple JSON REST services are used. This was deliberately 
+modelled on the implementation standards used in other Flanders Heritage sytems.
+This allows developers used to working in an enterprise IT context to feel 
+comfortable and productive.
 
 ![Searching for airfields in the Inventory of Immovable Cultural Heritage\label{fig:searchingairfields}](inventaris_screen_search_airfields.png)
 
-For external uses, publishing of linked data is supported. Individual concepts and
-collections can be downloaded as RDF data in Turtle, RDF/XML and JSON-LD format [@jsonld:2020].
-Entire conceptschemes can be downloaded in Turtle or RDF/XML format. Finally, 
-an [integration with a Linked Data Fragments (LDF) server](https://atramhasis.readthedocs.io/en/latest/development.html#running-a-linked-data-fragments-server) 
+While interactions with internal system are done through plain JSON REST
+services, publishing of linked data for external consumption is also supported.
+Individual concepts and collections can be downloaded as RDF data in Turtle, 
+RDF/XML and JSON-LD format [@jsonld:2020]. Entire conceptschemes can be downloaded 
+in Turtle or RDF/XML format. Finally, an 
+[integration with a Linked Data Fragments (LDF) server](https://atramhasis.readthedocs.io/en/latest/development.html#running-a-linked-data-fragments-server) 
 can be set up. An LDF server such as the 
 [Flanders Heritage Thesaurus LDF server](https://thesaurus.onroerenderfgoed.be/ldf/)
 can be browsed online for basic usage, but really shines in combination with an 
@@ -132,67 +134,73 @@ concepts from such a thesaurus and turning your vocabularies into true linked da
 
 # Technical requirements
 
-We decided to write a simple but extensible editor in Python that
-adhered to our primary and secondary requirements. Atramhasis comes with a 
-default style but is easy to extend with a custom corporate identity, as
-can be seen by comparing the [Flanders Heritage thesaurus](https://thesaurus.onroerenderfgoed.be) 
-with the default Atramhasis setup. It does not come with a default authentication 
-and a authorization layer, but the underlying Pyramid framework provides hooks
-and integration points facilitating this. There are default libraries for this 
-framework that can be configured according to a user's own corporate security 
-needs.
+When considering an IT-project, we take into account both functional and 
+non-functional requirements. These are less about what the software does, 
+and more about how it does it.
 
-Because the software uses [SQLAlchemy](https://sqlalchemy.org), a database 
[…]
