+++ paper/paper.bib
@@ -81,6 +81,20 @@
 	number = {196}
 }
 
+@inbook{Slechten:2004,
+	year = {2004},
+	title = {Namen noemen. Het CAI-thesaurusproject},
+	publisher = {Instituut voor het Archeologisch Patrimonium},
+	booktitle = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
+	volume = {14},
+	series = {I.A.P. Rapporten},
+	pages = {49--54},
+	url = {https://oar.onroerenderfgoed.be/publicaties/IAPR/14/IAPR014-003.pdf},
+	abstract = {Derde artikel van CAI-rapport 1 uit 2004, over de implementatie en het gebruik van een thesaurus in de CAI.},
+	keywords = {CAI, Centrale Archeologische Inventarisatie, archeologie, thesaurus},
+	author = {Katrien Slechten}
+}
+
 @misc{skos:2008,
   	abstract = {This document defines the Simple Knowledge Organization System (SKOS), a common data model for sharing and linking knowledge organization systems via the Web. Many knowledge organization systems, such as thesauri, taxonomies, classification schemes and subject heading systems, share a similar structure, and are used in similar applications. SKOS captures much of this similarity and makes it explicit, to enable data and technology sharing across diverse applications. The SKOS data model provides a standard, low-cost migration path for porting existing knowledge organization systems to the Semantic Web. SKOS also provides a light weight, intuitive language for developing and sharing new knowledge organization systems. It may be used on its own, or in combination with formal knowledge representation languages such as the Web Ontology language (OWL). This document is the normative specification of the Simple Knowledge Organization System. It is intended for readers who are involved in the design and implementation of information systems, and who already have a good understanding of Semantic Web technology, especially RDF and OWL.},
   	author = {Miles, Alistair and Bechhofer, Sean},
