### 2023-02-08 9a1f558
Clean bibliography
