+++ paper/paper.bib
@@ -63,7 +63,7 @@
 	year = 2021,
 	month = {sep},
 	institution = {agentschap Onroerend Erfgoed},
-	author = {Aukje de Haan and Nele Vanmaele},
+	author = {de Haan, Aukje and Vanmaele, Nele},
 	title = {Inventariseren van landschappelijk erfgoed in Vlaanderen. Historiek, methodologie, doelstellingen en resultaten},
 	series = {Onderzoeksrapporten agentschap Onroerend Erfgoed},
 	number = {195}
@@ -83,16 +83,17 @@
 
 @inbook{Slechten:2004,
 	year = {2004},
-	title = {Namen noemen. Het CAI-thesaurusproject},
+	editor = {Meylemans, Erwin}
+	title = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
+	chapter = {Namen noemen. Het CAI-thesaurusproject},
 	publisher = {Instituut voor het Archeologisch Patrimonium},
-	booktitle = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
 	volume = {14},
 	series = {I.A.P. Rapporten},
 	pages = {49--54},
 	url = {https://oar.onroerenderfgoed.be/publicaties/IAPR/14/IAPR014-003.pdf},
 	abstract = {Derde artikel van CAI-rapport 1 uit 2004, over de implementatie en het gebruik van een thesaurus in de CAI.},
 	keywords = {CAI, Centrale Archeologische Inventarisatie, archeologie, thesaurus},
-	author = {Katrien Slechten}
+	author = {Slechten, Katrien}
 }
 
 @misc{skos:2008,
