+++ paper/paper.bib
@@ -83,7 +83,7 @@
 
 @inbook{Slechten:2004,
 	year = {2004},
-	editor = {Meylemans, Erwin}
+	editor = {Meylemans, Erwin},
 	title = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
 	chapter = {Namen noemen. Het CAI-thesaurusproject},
 	publisher = {Instituut voor het Archeologisch Patrimonium},
