+++ paper/paper.bib
@@ -84,10 +84,10 @@
 @inbook{Slechten:2004,
 	year = {2004},
 	editor = {Meylemans, Erwin},
-	title = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
-	chapter = {Namen noemen. Het CAI-thesaurusproject},
+	booktitle = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
+	title = {Namen noemen. Het CAI-thesaurusproject},
 	publisher = {Instituut voor het Archeologisch Patrimonium},
-	volume = {14},
+	number = {14},
 	series = {I.A.P. Rapporten},
 	pages = {49--54},
 	url = {https://oar.onroerenderfgoed.be/publicaties/IAPR/14/IAPR014-003.pdf},
