+++ paper/paper.bib
@@ -84,7 +84,7 @@
 @inbook{Slechten:2004,
 	year = {2004},
 	editor = {Meylemans, Erwin},
-	booktitle = {CAI-I. De opbouw van een archeologisch beleidsinstrument},
+	booktitle = {{CAI-I.} De opbouw van een archeologisch beleidsinstrument},
 	title = {Namen noemen. Het CAI-thesaurusproject},
 	publisher = {Instituut voor het Archeologisch Patrimonium},
 	number = {14},
