### 2023-02-09 #5040 @koenedaele
@gaurav and @SvenLieber We have prepared a new draft, addressing your requests regarding the paper itself. Could you please review this new draft?

Tickets have been created for some of the questions about software functionality, but I still need to address some points made by @SvenLieber
