### 2023-02-23 #5040 @koenedaele
I'm not familiar with the bot, so I'm not sure why it's not working. I can access the paper through the Github actions log: https://github.com/OnroerendErfgoed/atramhasis/actions/runs/4132538083 but that doesn't work when I'm logged out.
