### 2023-02-23 #5040 @koenedaele
Seems to be a difference between the joss bot and the Github action I tested with regarding affiliations.
