### 2023-02-24 #5040 @gaurav
Thanks @koenedaele and @SvenLieber for figuring out the PDF issues! I've [updated my checklist](https://github.com/openjournals/joss-reviews/issues/5040#issuecomment-1372469208). I really like the new manuscript with all those screenshots and all the new details!

I am a little concerned that the manuscript may now be too long ([JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) state that "Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper. [...]
A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience. [...] Given this format, a “full length” paper is not permitted [...]"). I personally think this level of detail is fine, but I'd defer to @danielskatz's opinion on what is acceptable for JOSS. If you do need to remove some words to meet the word limit, I would suggest moving some of the technical details (e.g. PostgreSQL vs SQLite, skosproviders) into the repository README file.

Some minor suggested fixes:
- [ ] On line 65 (page 3), "Atramhasis" is misspelled as "Atramahsis".
- [ ] On line 71 (page 5), "possibility" is misspelled as "possiblity".
- [ ] On line 75 (page 5), "it's" should be "its".
- [ ] On line 96 (page 5), "multiple" is misspelled "mutiple".
- [ ] On line 101 (page 5), I wouldn't mention "While this previously needed to be done by writing a small piece of code, version 2.0.0 will provide a web interface for this.", since version 2.0.0 hasn't been released yet. I think it's fine to leave this out of the manuscript and include it in the repository documentation.

Unrelated to the manuscript (so you might still be working on this), I would strongly recommend that you provide a brief overview of the submission process to the repository README file or, at a minimum, provide a link to [CONTRIBUTING.md](https://github.com/OnroerendErfgoed/atramhasis/blob/develop/CONTRIBUTING.md) in that README file.

Apart from that, all my remaining comments are to do with the "Functionality" part of your software. I'll try running it this weekend (unless there's something there you're still working on) and report back on how that goes!
