### 2023-02-24 #5040 @koenedaele
Thank you. I'll fix those minor issues and update the README, it's indeed rather sparse at the moment.

I did indeed notice that we pushed the word limit. In the first version I tried to stick close to the word limit, but due to all the interesting questions that arose during the review I had to increase the word count. Atramhasis is a verry flexible system that can be used in lots of different ways, but that does require some explaining. I'll await further feedback from @SvenLieber and @danielskatz before making changes to address this.

> On line 101 (page 5), I wouldn't mention "While this previously needed to be done by writing a small piece of code, version 2.0.0 will provide a web interface for this.", since version 2.0.0 hasn't been released yet. I think it's fine to leave this out of the manuscript and include it in the repository documentation.

I added this to address a question by @SvenLieber, I'm okay with removing it again if this is ok for him.
