### 2023-02-24 #5040 @koenedaele
As to the functionality, we are working on version 2.0.0 (https://github.com/OnroerendErfgoed/atramhasis/issues/735). That issue contains all the new things we're working on (we do use the ZenHub plugin and I'm not sure if the associated issues show up if you don't have Zenhub). There are some minor improvements that come from @SvenLieber 's review that are already included, feel free to add more. I've move his comments on the functionality to https://github.com/OnroerendErfgoed/atramhasis/issues/772 to make it easier to follow and discuss, feel free to chime in. The 2.0.0 has two major new functionalities:

- Allow non-numeric id's for conceptschemes (OnroerendErfgoed/atramhasis#732)
- Allow configuring new conceptschems through the UI (OnroerendErfgoed/atramhasis#724)
