+++ README.rst
@@ -18,5 +18,17 @@ Atramhasis is an online SKOS editor. This webapplication enables
 users to create SKOS vocabularies consisting of Concepts and Collections. It 
 tries to stay close to the SKOS specification. This makes it suited for different
 types of vocabularies such as simple pick lists, authority files, flat lists and
-basic to relatively complex thesauri. Atramhasis is not suited for managing very
-large or complex thesauri such as the `AAT <http://vocab.getty.edu>`_. 
+basic to relatively complex thesauri. For a full overview of capabilities, please
+consult our `documentation <https://atramhasis.readthedocs.io/en/latest/features.html>`_. 
+If you want a quick demo of what the system provides for a typical end user, 
+have a look at the `Flanders Heritage Thesaurus <https://thesaurus.onroerenderfgoed.be>`_
+
+All development is done through `Github <https://github.com/OnroerendErfgoed/atramhasis>`_. 
+If you run into bugs or would like to request a new feature, please open a 
+`Github issue <https://github.com/OnroerendErfgoed/atramhasis/issues>`_. Please
+provide some context to your question such as the operating system,
+Python and Atramhasis versions you are running. If you have specific questions
+about the software or the datasets hosted at
+`Flanders Heritage Thesaurus <https://thesaurus.onroerenderfgoed.be>`_
+you would rather not address in a public forum, please mail us at
+ict@onroerenderfgoed.be.
