+++ README.rst
@@ -21,7 +21,15 @@ types of vocabularies such as simple pick lists, authority files, flat lists and
 basic to relatively complex thesauri. For a full overview of capabilities, please
 consult our `documentation <https://atramhasis.readthedocs.io/en/latest/features.html>`_. 
 If you want a quick demo of what the system provides for a typical end user, 
-have a look at the `Flanders Heritage Thesaurus <https://thesaurus.onroerenderfgoed.be>`_
+have a look at the `Flanders Heritage Thesaurus <https://thesaurus.onroerenderfgoed.be>`_. 
+This is an implementation of Atramhasis for the `Flanders Heritage Agency <https://www.onroerenderfgoed.be>`_
+containing several vocabularies regarding cultural heritage, such as 
+`heritagetypes <https://thesaurus.onroerenderfgoed.be/conceptschemes/ERFGOEDTYPES>`_, 
+`styles and cultures <https://thesaurus.onroerenderfgoed.be/conceptschemes/STIJLEN_EN_CULTUREN>`_, 
+`materials <https://thesaurus.onroerenderfgoed.be/conceptschemes/MATERIALEN>`_ 
+and `periods <https://thesaurus.onroerenderfgoed.be/conceptschemes/DATERINGEN`_.
+Most of the vocabularies are in Dutch, but the interface is available in both 
+Dutch, English and French.
 
 All development is done through `Github <https://github.com/OnroerendErfgoed/atramhasis>`_. 
 If you run into bugs or would like to request a new feature, please open a 
@@ -33,6 +41,7 @@ about the software or the datasets hosted at
 you would rather not address in a public forum, please mail us at
 ict@onroerenderfgoed.be.
 
-Contributions are welcome, both in code and documentation. Please see our 
+Contributions are welcome. This could be improvements to the code, updated 
+documentation or adding a new translation to the User Interface. Please see our 
 `contribution guidelines <https://github.com/OnroerendErfgoed/atramhasis/blob/develop/CONTRIBUTING.md>`_ 
 to find out how we can collaborate.
