### 2023-02-24 1f7d682
Fix scrutinizer config.
