### 2023-02-24 1f6ef36
Fix scrutinizer config.
