### 2023-02-24 42aeec6
Exclude scripts from code analysis.
