### 2023-02-24 7e19bda
Exclude fixtures from code analysis.
