+++ atramhasis/views/views.py
@@ -70,7 +70,7 @@ class AtramhasisView:
     @view_config(name='favicon.ico')
     def favicon_view(self):
         """
-        This view returns the favicon when requested from the web root.
+        Return the favicon when requested from the web root.
         """
         here = os.path.dirname(__file__)
         icon = os.path.join(os.path.dirname(here), 'static', 'img', 'favicon.ico')
@@ -84,7 +84,7 @@ class AtramhasisView:
     @view_config(route_name='home', renderer='atramhasis:templates/home.jinja2')
     def home_view(self):
         """
-        This view displays the homepage.
+        Display the homepage.
         """
         conceptschemes = [
             {'id': x.get_metadata()['id'],
@@ -98,7 +98,7 @@ class AtramhasisView:
     @view_config(route_name='conceptschemes', renderer='atramhasis:templates/conceptschemes.jinja2')
     def conceptschemes_view(self):
         """
-        This view displays a list of available conceptschemes.
+        Display a list of available conceptschemes.
         """
         conceptschemes = [
             {'id': x.get_metadata()['id'],
@@ -113,7 +113,7 @@ class AtramhasisView:
     @view_config(route_name='conceptscheme', renderer='atramhasis:templates/conceptscheme.jinja2')
     def conceptscheme_view(self):
         """
-        This view displays conceptscheme details.
+        Display a single conceptscheme.
         """
         conceptschemes = [
             {'id': x.get_metadata()['id'],
@@ -149,7 +149,7 @@ class AtramhasisView:
     @view_config(route_name='concept', renderer='atramhasis:templates/concept.jinja2')
     def concept_view(self):
         """
-        This view displays the concept details
+        Display all about a single concept or collection.
         """
         conceptschemes = [
             {'id': x.get_metadata()['id'],
@@ -189,7 +189,7 @@ class AtramhasisView:
     @view_config(route_name='search_result', renderer='atramhasis:templates/search_result.jinja2')
     def search_result(self):
         """
-        This view displays the search results
+        Display search results
         """
         conceptschemes = [
             {'id': x.get_metadata()['id'],
@@ -231,7 +231,7 @@ class AtramhasisView:
     @view_config(route_name='locale')
     def set_locale_cookie(self):
         """
-        This view will set a language cookie
+        Set a language cookie to remember what language a user requested.
         """
         settings = get_current_registry().settings
         default_lang = settings.get('pyramid.default_locale_name')
@@ -255,6 +255,9 @@ class AtramhasisView:
     @audit
     @view_config(route_name='search_result_export', renderer='csv')
     def results_csv(self):
+        """
+        Download search results in CSV format, allowing further processing.
+        """
         header = ['conceptscheme', 'id', 'uri', 'type', 'label', 'prefLabels', 'altLabels', 'definition', 'broader',
                   'narrower', 'related']
         rows = []
