### 2023-02-24 710d684
Refactor method to get public conceptschemes. Refs #777 (#778)
* Refactor method to get public conceptschemes. Refs #777

* Update changelog. Refs #777
