### 2023-02-24 f199065
Turn method into function. Refs #777
