### 2023-02-24 2768388
Further tests for get_public_conceptschemes. Refs #777
