### 2023-02-24 4ad154a
Fix tests for get_public_conceptschemes. Refs #777
