### 2023-02-28 #5040 @SvenLieber
Thank you for the revised paper (and addressing other opened issues).
I like the new version of the paper and I think that it is an adequate and very informative description of the software. One that not only describes the features well, but also provides information around design decisions and the state of the art, very relevant for individuals or organizations who wish to use/contribute to the software.

> On line 101 (page 5), I wouldn't mention "While this previously needed to be done by writing a small piece of code, version 2.0.0 will provide a web interface for this.", since version 2.0.0 hasn't been released yet. I think it's fine to leave this out of the manuscript and include it in the repository documentation.

It's fine for me to remove that part @koenedaele .

I also agree with @gaurav that some details such as the information around the SQL implementation can be shortened if this is needed.

As for the checklist @danielskatz, I went over it and updated it where necessary.
I am happy with all changes and consider the checklist fulfilled.

Of course I am available if there are more questions or updates on the paper which would require my attention.
