### 2023-02-28 #5040 @gaurav
> @gaurav - can you examine this again in terms of the items you haven't checked off? And let us know what else is needed from the authors?

My apologies: I hadn't noticed that the [documentation for starting a demo instance](https://atramhasis.readthedocs.io/en/latest/demo.html) has been updated since I last looked at it. I followed the new instructions and it started with only one minor problem (https://github.com/OnroerendErfgoed/atramhasis/issues/779). I'd like a day or two to make sure that "the functional claims of the software been confirmed" -- that should be the last remaining item on my checklist!
