### 2023-02-28 OnroerendErfgoed/atramhasis#779 @koenedaele
Last time I edited the demo instructions I had no issues (Ubuntu Linux). I just checked and pytz is present in that venv. Is it possible you're using MacOSX (based on the home dir locations) and this issue is OS related?
