### 2023-03-01 OnroerendErfgoed/atramhasis#779 @koenedaele
pytz should come installed with Babel, but apparently it isn't. There's actually no reason for pytz to be in that delete_scheme script, so it can be removed there. It does have a use in the create_sitemap script, where it seems we've hardcoded the timezone to ours, so that could be done better.
