### 2023-03-01 #5040 @koenedaele
I've removed the reference to version 2. That concludes the latest checklist from @gaurav
