### 2023-03-01 #5040 @koenedaele
I've also shortened the SQL paragraph. More or less the same information, but less verbose.
