### 2023-03-01 336b537
Remove pytz from delete_scheme script (#782)
* Remove pytz from delete_scheme script. Refs #780

* Update changelog. Refs #780
