### 2023-03-02 e7f2648
Update jina2 versions. Refs #747
