### 2023-03-02 6d7c67b
Feature/747 jinja3 (#784)
* Update jina2 versions. Refs #747

* Remove unneeded with assignments. Refs #747

* Update changelog. Refs #747

* Update copyright in static admin. Refs #747

* Remove unneeded copyright sets. Refs #747

* fix typo

* Remove old IE 9 code. Refs #747

* Remove unneeded jinja2 template. Refs #747

---------

Co-authored-by: Vanderhaegen Cedrik <cedrik.vanderhaegen@vlaanderen.be>
