+++ requirements.txt
@@ -20,7 +20,7 @@ zope.sqlalchemy==1.6
 transaction==3.0.1
 
 # jinja2
-jinja2==2.11.3
+jinja2==3.1.2
 markupsafe==2.0.1
 pyramid-jinja2==2.10
 Babel==2.11.0
