### 2023-03-04 OnroerendErfgoed/atramhasis#794 @gaurav
JSON-LD "@context" needs another "@context"
Steps to reproduce:
1. Go to any concept, e.g. https://thesaurus.onroerenderfgoed.be/conceptschemes/ERFGOEDTYPES/c/885
2. Click "JSON/LD" to download the JSON-LD file as `885.jsonld`.
3. Run `npx jsonld-cli format -q 885.jsonld` to convert the JSON-LD file into N-Quads.

Found:
* No output.

Expected:
* N-Quads output.

Fix:
1. Download https://thesaurus.onroerenderfgoed.be/jsonld/context/skos and save it as `skos.json`, e.g. `wget "https://thesaurus.onroerenderfgoed.be/jsonld/context/skos" -O skos.json`
2. Edit skos.json by adding an additional `{"@context":{...}}` around the entire contents of the file ([as in this file](https://github.com/OnroerendErfgoed/atramhasis/files/10887096/skos.json.txt)) (I got this idea from this StackExchange answer: https://stackoverflow.com/a/61231984/27310)
3. In 885.jsonld, change the "@context" from "https://thesaurus.onroerenderfgoed.be/jsonld/context/skos" to "./skos.json" ([as in this example](https://github.com/OnroerendErfgoed/atramhasis/files/10887106/885.jsonld.txt)).
4. Run `npx jsonld-cli format -q 885.jsonld`. Valid N-quads are now produced.

So: I think you need to change the code that generates `/jsonld/context/skos` so that it adds `{"@context":{...}}` around the current output for it to be a functional JSON-LD context.
