### 2023-03-04 #5040 @gaurav
Hi everybody, thank you for giving me the opportunity to spend a few days diving deeply into Atramhasis! As I mentioned, I was able to set up a demo locally by following the instructions at https://atramhasis.readthedocs.io/en/latest/demo.html, and as far as I can tell this instance works pretty well! I filed a couple of minor issues in the Atramhasis repository, but these are generally not major and I'd be happy to accept this manuscript without them. I was also pleased to see that after installing Tox with `pip install tox`, I was able to run the test suite with 255 successful tests -- good job!

However, there are two issues I'd like to see addressed before I sign off on this review:
1. AFAICT the JSON-LD files downloaded from Atramhasis don't work. I was able to fix them by adding an additional "@context" around the contents of the context file (https://github.com/OnroerendErfgoed/atramhasis/issues/794). Since you mention the JSON-LD export in the manuscript, I would like to confirm that Atramhasis has working JSON-LD files before checking off the "Functionality: Have the functional claims of the software been confirmed?" item. **Please** do let me know if I've just done something wrong, and if there is some tool for which this JSON-LD currently works properly!
2. I'm not sure what the purpose of Figure 3 (line 48, page 3) is in this manuscript -- I had assumed it was a screenshot of Atramhasis, but on trying to replicate it to day I noticed that it's actually an unrelated tool. I would suggest either removing it or explaining how Atramhasis is involved in the functionality being displayed.

Thanks again, and I think these should the last two requests from me!
