### 2023-03-06 OnroerendErfgoed/atramhasis#794 @koenedaele
Yes. Good catch. Thank you!
