+++ paper/paper.bib
@@ -57,6 +57,17 @@
   	url       = {https://comunica.github.io/Article-ISWC2018-Resource/}
 }
 
+@article{Fernandez:2013,
+	author = {Fernández, Javier D. and Martínez-Prieto, Miguel A. and Gutiérrez, Claudio and Polleres, Axel and Arias, Mario},
+	title = {Binary RDF Representation for Publication and Exchange (HDT)},
+	doi = {10.2139/ssrn.3198999}
+	journal={Web Semantics: Science, Services and Agents on the World Wide Web},
+	volume={19},
+	pages={22–41},
+	year={2013},
+	publisher={Elsevier}
+}
+
 @techreport{deHaan:2021,
 	doi = {10.55465/rsdd4339},
 	url = {https://doi.org/10.55465/rsdd4339},
