+++ paper/paper.bib
@@ -1,3 +1,59 @@
+@software{skosprovider:2022,
+  author       = {Van Daele, Koen and
+                  Geuens, Jonas},
+  title        = {Skosprovider},
+  month        = aug,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {1.2.0},
+  doi          = {10.5281/zenodo.6984378},
+  url          = {https://doi.org/10.5281/zenodo.6984378}
+}
+
+@techreport{deHaan:2021,
+	doi = {10.55465/rsdd4339},
+	url = {https://doi.org/10.55465/rsdd4339},
+	year = 2021,
+	month = {sep},
+	institution = {agentschap Onroerend Erfgoed},
+	author = {de Haan, Aukje and Vanmaele, Nele},
+	title = {Inventariseren van landschappelijk erfgoed in Vlaanderen. Historiek, methodologie, doelstellingen en resultaten},
+	series = {Onderzoeksrapporten agentschap Onroerend Erfgoed},
+	number = {195}
+}
+
+@techreport{Hooft:2021,
+	doi = {10.55465/mmym4330},
+	url = {https://doi.org/10.55465/mmym4330},
+	year = 2021,
+	month = {sep},
+	institution = {agentschap Onroerend Erfgoed},
+	author = {Elise Hooft},
+	title = {Inventariseren van bouwkundig erfgoed in Vlaanderen. Historiek, methodologie, doelstellingen en resultaten},
+	series = {Onderzoeksrapporten agentschap Onroerend Erfgoed},
+	number = {196}
+}
+
+@misc{jsonld:2020,
+	abstract = {JSON is a useful data serialization and messaging format. This specification defines JSON-LD 1.1, a JSON-based format to serialize Linked Data. The syntax is designed to easily integrate into deployed systems that already use JSON, and provides a smooth upgrade path from JSON to JSON-LD. It is primarily intended to be a way to use Linked Data in Web-based programming environments, to build interoperable Web services, and to store Linked Data in JSON-based storage engines.},
+	editor = {Kellogg, Greg and Champin, Pierre-Antoine and Longley, Dave},
+	url = {https://www.w3.org/TR/2020/REC-json-ld11-20200716/},
+	title = {{JSON-LD 1.1}},
+	year = 2020,
+	month = {July},
+  	series = {W3C Recommendation},
+  	publisher = {World Wide Web Consortium}
+}
+
+@inproceedings{Taelman:2018,
+  	author    = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
+  	title     = {Comunica: a Modular SPARQL Query Engine for the Web},
+  	booktitle = {Proceedings of the 17th International Semantic Web Conference},
+  	year      = {2018},
+  	month     = oct,
+  	url       = {https://comunica.github.io/Article-ISWC2018-Resource/}
+}
+
 @article{Mortier:2017,
 	title = {Calling it what it is. {Thesauri} in the {Flanders} {Heritage} {Agency}: {History}, {Importance}, {Use} and {Technological} {Advances}},
 	volume = {IV-2/W2},
@@ -38,6 +94,28 @@
 	pages = {323--327},
 }
 
+@misc{rdf:2014,
+	abstract = {The Resource Description Framework (RDF) is a language for representing information about resources in the World Wide Web. This Primer is designed to provide the reader with the basic knowledge required to effectively use RDF. It introduces the basic concepts of RDF and describes its XML syntax. It describes how to define RDF vocabularies using the RDF Vocabulary Description Language, and gives an overview of some deployed RDF applications. It also describes the content and purpose of other RDF specification documents.},
[…]
