### 2023-03-08 OnroerendErfgoed/atramhasis#794 @koenedaele
Todo: release pyramid_skosprovider 1.2.0 and update in Atramhasis.
