### 2023-03-08 OnroerendErfgoed/atramhasis#755 @Wim-De-Clercq
Extra things i'll fix here:

- admin interface having css issues
- there's an unnecessary __init__.py in the root.
