### 2023-03-08 28b2142
Update pyramid_skosprovider. Refs #794
