### 2023-03-08 f7c3b37
Update pyramid_skosprovider. Refs #794 (#797)
