### 2023-03-08 24c26ef
Update pyramid_skosprovider. Refs #794 (#797)
