+++ CHANGES.rst
@@ -15,7 +15,7 @@ in progress
 - Stop building universal wheels (#752)
 - Update docs to no longer reference mkvirtualenv, use standard venv instead (#773)
 - Update CONTRIBUTING.md file (#756)
-- Remove unneeded reference to pytz (#780)
+- Remove unneeded references to pytz (#780, #796)
 - Minor refactoring of view method to get publically available conceptschemes (#777)
 
 1.3.0 (04-01-2023)
