### 2023-03-08 #5040 @koenedaele
The  JSON-LD issue has been fixed in the current development version. Since it comes from a library, it's possible to update the reference to this library in the demo environment being used to see the effect.
