### 2023-03-09 OnroerendErfgoed/atramhasis#759 @koenedaele
Documentation was updated. References to virtualenv have been removed. Instructions now include upgrading setuptools and pip.
