### 2023-03-09 70dbf1d
Add GET provider endpoints. (#795)
Use openapi to validate if we have it anyway.
Use separate adapter for rendering SQLAlchemyProvider,

Issue: #790
