### 2023-03-13 86e9bef
pass full schemes to search pane
