+++ CHANGES.rst
@@ -1,6 +1,8 @@
 1.3.1 (13-03-2023)
 ------------------
 
+- Show acutal language tags in admin interface drop downs, as opposed to just the labels to reduce confusion. (#766)
+- Update docs to no longer reference mkvirtualenv, use standard venv instead (#773)
 - Update CONTRIBUTING.md file (#756)
 
 1.3.0 (04-01-2023)
