### 2023-03-14 #5040 @koenedaele
@danielskatz Currently the paper resides in a branch that hasn't been merged with the default branch. Should I merge the paper into the version that will be archived on Zenodo?
