### 2023-03-14 d9b81ae
Handle uri_generator being None for rendering provider.
Issue #790
