+++ tests/test_views.py
@@ -4,25 +4,35 @@ from unittest.mock import Mock
 
 from paste.deploy.loadwsgi import appconfig
 from pyramid import testing
+from pyramid.config.settings import Settings
+from pyramid.request import apply_request_extensions
+from pyramid.testing import DummyRequest
 from skosprovider.registry import Registry
+from skosprovider.uri import DefaultUrnGenerator
 from skosprovider_sqlalchemy.models import Collection
 from skosprovider_sqlalchemy.models import Concept
+from skosprovider_sqlalchemy.models import ConceptScheme
 from skosprovider_sqlalchemy.models import Label
 from skosprovider_sqlalchemy.models import LabelType
 from skosprovider_sqlalchemy.models import Note
 from skosprovider_sqlalchemy.models import Thing
+from skosprovider_sqlalchemy.providers import SQLAlchemyProvider
 from sqlalchemy.exc import NoResultFound
 from webob.multidict import MultiDict
+from skosprovider.skos import Concept as SkosConcept
 
+import atramhasis
 from atramhasis.errors import ConceptNotFoundException
 from atramhasis.errors import ConceptSchemeNotFoundException
 from atramhasis.errors import SkosRegistryNotFoundException
+from atramhasis.skos import IDGenerationStrategy
+from atramhasis.views.crud import AtramhasisCrud
 from atramhasis.views.views import AtramhasisAdminView
 from atramhasis.views.views import AtramhasisListView
 from atramhasis.views.views import AtramhasisView
 from atramhasis.views.views import get_definition
-from atramhasis.views.views import labels_to_string
 from atramhasis.views.views import get_public_conceptschemes
+from atramhasis.views.views import labels_to_string
 from fixtures.data import trees
 
 TEST_DIR = os.path.dirname(__file__)
@@ -47,12 +57,14 @@ def hidden_provider(some_id):
     )
     return provider_mock
 
+
 def external_provider(some_id):
     provider_mock = provider(some_id)
     provider_mock.get_metadata = Mock(
         return_value={"id": some_id, "subject": ["external"]}
     )
     return provider_mock
+
 
 class DummySKOSManager:
     def get_thing(self, concept_id, conceptscheme_id):
@@ -581,6 +593,7 @@ class TestViewFunctions(unittest.TestCase):
         conceptschemes = get_public_conceptschemes(regis)
         self.assertEqual(1, len(conceptschemes))
 
+
 class TestListViews(unittest.TestCase):
     def setUp(self):
         self.config = testing.setUp()
@@ -602,3 +615,98 @@ class TestListViews(unittest.TestCase):
         labellist = atramhasis_list_view.labeltype_list_view()
         self.assertIsNotNone(labellist)
         self.assertIn({"key": "prefLabel", "label": "prefLabel"}, labellist)
+
+
+class TestAtramhasisCrudView(unittest.TestCase):
+    class SKOSRegistry:
+        pass
+
+    class SKOSManager:
+        def get_all_label_types(self):
+            return [LabelType(name='prefLabel', description=''),
+                    LabelType(name='sortLabel', description='')]
+
+        def get_next_cid(self, _, __):
+            return "next_cid"
+
+        def save(self, concept):
+            if concept.id is None:
+                concept.id = 1
+            concept.conceptscheme = ConceptScheme(uri='urn:x-skosprovider:test')
+            return concept
+
+        def get_thing(self, concept_id, scheme_id):
+            raise NoResultFound()
+
+    class LanguagesManager:
+        def count_languages(self, _):
+            return 1
+
+    class Provider(SQLAlchemyProvider):
+        metadata = {'subject': 'stub'}
+        conceptscheme_id = 'cs_id'
[…]
