+++ CITATION.cff
@@ -88,4 +88,4 @@ keywords:
   - python
 license: GPL-3.0
 version: 1.3.1
-date-released: '2023-03-??'
+date-released: '2023-03-14'
