### 2023-03-14 294e396
Release/1.3.1 (#803)
* Update contributing.md file. Refs #756 (#757)

* Update contributing
* Update changelog

* #766 add extra id field to language dropdown

* rename param

* move property to options

* Remove dependency in docs on mkvirtualenv. Refs #773 (#775)

* Update README

* Add link to contibution guidelines.

* Update README.

* Update CHANGES.

* Remove pytz from delete_scheme script (#782)

* Remove pytz from delete_scheme script. Refs #780

* Update changelog. Refs #780

* trigger reload list on save

* Update pyramid_skosprovider. Refs #794 (#797)

* Update changelog.

* Drop p
[…]
