+++ requirements.txt
@@ -8,7 +8,7 @@ openapi-spec-validator==0.4.0  # https://github.com/p1c2u/openapi-core/issues/44
 # skosprovider
 skosprovider==1.2.0
 skosprovider_sqlalchemy==1.0.0
-pyramid_skosprovider==1.1.0
+pyramid_skosprovider==1.2.0
 skosprovider_rdf==1.3.0
 skosprovider_getty==1.1.0
 
