### 2023-03-14 1ac5b7b
Fix error in CITATION.cff
