+++ CITATION.cff
@@ -26,6 +26,7 @@ authors:
   - given-names: Bram
     family-names: Goessens
     affiliation: Agentschap Onroerend Erfgoed
+    orcid: 'https://orcid.org/0000-0001-6693-0866'
   - given-names: Jonathan
     family-names: Piraux
     affiliation: GEO Solutions
