+++ CITATION.cff
@@ -67,7 +67,7 @@ references:
       - given-names: Leen
         family-names: Meganck
         email: leen.meganck@vlaanderen.be
-     	affiliation: Agentschap Onroerend Erfgoed
+        affiliation: Agentschap Onroerend Erfgoed
         orcid: 'https://orcid.org/0000-0002-9483-0851'
     journal: ISPRS Annals of Photogrammetry, Remote Sensing and Spatial Information Sciences
     volume: 4W2
