+++ docs/source/conf.py
@@ -58,7 +58,7 @@ copyright = '2014-2023, Flanders Heritage Agency'
 # The short X.Y version.
 version = '1.3'
 # The full version, including alpha/beta/rc tags.
-release = '1.3.1'
+release = '1.3.2'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
