+++ CITATION.cff
@@ -88,5 +88,5 @@ keywords:
   - vocabulary
   - python
 license: GPL-3.0
-version: 1.3.1
+version: 1.3.2
 date-released: '2023-03-14'
