### 2023-03-14 #5040 @koenedaele
@danielskatz I've released this version as 1.3.2, with DOI 10.5281/zenodo.7733994 (https://zenodo.org/record/7733994)

Version 1.3.2 is actually identical to 1.3.1, apart from a formatting bug present in the CITATION.cff file that prevented Zenodo from parsing the 1.3.1 version.
