+++ paper/paper.bib
@@ -46,12 +46,19 @@
 }
 
 @inproceedings{Taelman:2018,
-  	author    = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
-  	title     = {Comunica: a Modular SPARQL Query Engine for the Web},
+  	author = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
+  	title = {Comunica: a Modular SPARQL Query Engine for the Web},
   	booktitle = {Proceedings of the 17th International Semantic Web Conference},
-  	year      = {2018},
-  	month     = oct,
-  	url       = {https://comunica.github.io/Article-ISWC2018-Resource/}
+  	year = {2018},
+  	month = oct,
+	series = {Lecture Notes in Computer Science},
+        volume = 11137,
+        editor = {Vrande\v{c}i\'{c}, Denny and Bontcheva, Kalina and Su\'arez-Figueroa, Mari Carmen and Presutti, Valentina and Celino, Irene and Sabou, Marta and Kaffee, Lucie-Aim\'ee and Simperl, Elena},
+        publisher = {Springer},
+        pages = {239--255},
+        isbn = {978-3-030-00668-6},
+  	url = {https://comunica.github.io/Article-ISWC2018-Resource/},
+	doi = {10.1007/978-3-030-00668-6_15},
 }
 
 @article{Mortier:2017,
