+++ paper/paper.bib
@@ -48,7 +48,7 @@
 @inproceedings{Taelman:2018,
   	author = {Taelman, Ruben and Van Herwegen, Joachim and Vander Sande, Miel and Verborgh, Ruben},
   	title = {Comunica: a Modular SPARQL Query Engine for the Web},
-  	booktitle = {Proceedings of the 17th International Semantic Web Conference},
+  	booktitle = {Proceedings of the 17th {I}nternational {S}emantic {W}eb {C}onference},
   	year = {2018},
   	month = oct,
 	series = {Lecture Notes in Computer Science},
