### 2023-03-14 #5040 @danielskatz
@koenedaele - I've made some suggestions for small changes in https://github.com/OnroerendErfgoed/atramhasis/pull/804 - please merge this or let me know what you disagree with, then we can continue the process of acceptance and publication.
