### 2023-03-14 cb95a1e
minor changes in paper/bib (#804)
* minor changes in paper

* minor change in bib
