+++ paper/paper.md
@@ -49,20 +49,20 @@ Semantic Web. SKOS provides a standard way to represent these vocabularies using
 passing data between computer applications in an interoperable way.  
 
 SKOS vocabularies record controlled vocabularies as a set of concepts, collections
-and their relations. A concept being something a researcher wants to describe and define,
-a collection being a grouping of a number of concepts.
+and their relations. A concept is something a researcher wants to describe and define,
+and a collection is a grouping of a number of concepts.
 Flanders Heritage is active in the domain of cultural heritage and 
 describes concepts such as [Roman period](https://id.erfgoed.net/thesauri/dateringen/1223) [@Slechten:2004],
-[oppida](https://id.erfgoed.net/thesauri/erfgoedtypes/1052) or 
+[oppida](https://id.erfgoed.net/thesauri/erfgoedtypes/1052), and 
 [archaeological excavations](https://id.erfgoed.net/thesauri/gebeurtenistypes/38). A 
 typical collection would be [settlements by function](https://id.erfgoed.net/thesauri/erfgoedtypes/1034), 
 a grouping of settlement types according to their function, as opposed to by size or form. 
-Similar concepts and collections are grouped in a conceptscheme, eg. the 
-conceptscheme of [heritage types](https://id.erfgoed.net/thesauri/erfgoedtypes)
-consists of concepts and collections that describes types of heritage, ranging from 
+Similar concepts and collections are grouped in a conceptscheme, e.g., the 
+[heritage types](https://id.erfgoed.net/thesauri/erfgoedtypes) conceptscheme
+consists of concepts and collections that describe types of heritage, ranging from 
 [solitary trees](https://id.erfgoed.net/thesauri/erfgoedtypes/1654) 
 over [burial mounds](https://id.erfgoed.net/thesauri/erfgoedtypes/170)
- to [airfields](https://id.erfgoed.net/thesauri/erfgoedtypes/476) 
+to [airfields](https://id.erfgoed.net/thesauri/erfgoedtypes/476) 
 and [swimming pools](https://id.erfgoed.net/thesauri/erfgoedtypes/949). Concepts 
 and collections can be labelled with preferred labels and alternative labels, 
 more elaborately described with definitions and notes, and provided with references 
@@ -78,8 +78,8 @@ to concepts in other vocabularies such as the Art and Architecture Thesaurus (AA
 
 Atramhasis was written to be a user-friendly open source SKOS editor. First and 
 foremost, we wanted a system that adheres to the SKOS standard, yet is useable 
-for users without knowledge of SKOS or RDF. For a typical user, it had to feel 
-as if they were consulting a normal website, as opposed to a RDF vocabulary 
+for users without prior knowledge of SKOS or RDF. For a typical user, it had to feel 
+as if they were consulting a normal website, as opposed to a RDF vocabulary, 
 since the latter can feel rather daunting to non-technical users. This not only 
 applies to users consulting the thesauri, but also to those editing them. Our 
 thesaurus editors are not IT- or linked data specialists, but domain experts, 
@@ -87,14 +87,14 @@ researchers and other specialists in the field of cultural heritage. While a
 general understanding of thesauri is within their grasp, the technicalities of 
 RDF are not. Thus, editors in Atramhasis do not write RDF statements, but edit 
 data in a normal web admin interface, as seen in \autoref{fig:editingairfields}. 
-All mapping to RDF and SKOS is done behind the scene, invisible to the editors.
+All mapping to RDF and SKOS is done behind the scenes, invisible to the editors.
 
 ![Editing the concept of airfields is simple and straightforward.\label{fig:editingairfields}](atramhasis_screen_edit_airfields.png)
 
[…]
